[AL-18] refactor. refactoring directory structure
refactoring directory structure like @sezikim
