+++ src/main/java/com/progm/allsinsa/cart/controller/CartProductController.java
@@ -1,6 +1,7 @@
-package com.progm.allsinsa.cart.cartProduct;
+package com.progm.allsinsa.cart.controller;
 
-import com.progm.allsinsa.cart.cart.CartService;
+import com.progm.allsinsa.cart.dto.CartProductDto;
+import com.progm.allsinsa.cart.service.CartProductService;
 import javassist.NotFoundException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
