+++ src/test/java/com/progm/allsinsa/cart/cart/CartControllerTest.java
@@ -1,9 +1,7 @@
 package com.progm.allsinsa.cart.cart;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.progm.allsinsa.cart.CartService;
+import com.progm.allsinsa.cart.service.CartService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
